requirements.txt
pandas==2.1.0
sqlalchemy==2.0.25
pyarrow==14.0.1
pymysql==1.1.0
psycopg2-binary==2.9.9
pyyaml==6.0
numpy==1.26.0
pytest==7.4.4